+++ src/commonMain/kotlin/com/insanusmokrassar/sdi/HelperTools.kt
@@ -1,0 +1,7 @@
+package com.insanusmokrassar.sdi
+
+import kotlinx.serialization.ImplicitReflectionSerializer
+import kotlinx.serialization.json.Json
+
+@ImplicitReflectionSerializer
+fun loadModule(json: String): Module = Json.nonstrict.parse(Module.serializer(), json)