+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.1.2
 
+* All reflection support methods are internal
+
 ## 0.1.1
 
 * Added opportunity to create objects inside of config: