+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * All reflection support methods are internal
 * Added `loadModule` functions for more useful working with modules loading
-* `Module` now is not serializable and can be created only via `loadModule` functions
+* `Module` now is not serializable and can be created only via `loadModule` functions (at least, for some time)
 
 ## 0.1.1